backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
motor==3.1.1
pydantic==1.10.13
python-jose==3.3.0
passlib==1.7.4